requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# AI and LLM
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.5

# MCP Protocol
websockets>=11.0.0

# Testing
pytest>=7.4.0